requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.0.1
scikit-learn==1.3.2
tensorflow-cpu==2.17.0
